feat: 칼라 개발 관리 추가 - 변경사항 : 칼라 개발 관리 화면 개발 완료  - 칼라 개발 관리 조회, 칼라 개발 관리 등록 및 확인 화면 추가  - 개발접수부터 개발완료, 개발중단까지 총 8단계의 진도 관리가 가능하며, 순차 결재 시스템을 도입함  - 첨부파일 목록을 미리확인할 수 있도록 하고, 전체다운로드 기능 추가  - 특이사항 관리 기능, 접수담당자 변경, 부서변경, BOM 변경 기능 추가 - 변경사유 : 칼라제품 개발 이력과 관리 강화 - 참고 : 적용승인요청서 DKSYS-부산운영팀 2026-000811

+++ WebContents/_fileDeleteHandler6.jsp
@@ -49,18 +49,22 @@ else if("05".equals(IMG_RGS_FLAG))
 	FILE_ADDR = "/APP/WAS/FILES/C10/05" ;
 else if("06".equals(IMG_RGS_FLAG))  
 	FILE_ADDR = "/APP/WAS/FILES/C10/06" ;
-else if("07".equals(IMG_RGS_FLAG))  
+else if("07".equals(IMG_RGS_FLAG))
 	FILE_ADDR = "/APP/WAS/FILES/C10/07" ;
+else if("09".equals(IMG_RGS_FLAG))
+	FILE_ADDR = "/APP/WAS/FILES/C10/09" ;
+
+String SOFT_DELETE = request.getParameter("SOFT_DELETE");
 
 try{
-	File file = new File(FILE_ADDR , FILE_NAME);
-	
-	if(file.exists()){
-		file.delete();
+	// 소프트 삭제가 아닌 경우에만 실제 파일 삭제
+	if (!"Y".equals(SOFT_DELETE)) {
+		File file = new File(FILE_ADDR , FILE_NAME);
+		if(file.exists()){
+			file.delete();
+		}
 	}
-	
 
-	
 	PosContext ctx = new PosContext();
 	ctx.put("IMG_RGS_FLAG"		, IMG_RGS_FLAG);  //06
 	ctx.put("IMG_RGS_FLAG_ID"	, IMG_RGS_FLAG_ID);  //W4720
